refactor: clean up file names and imports
