+++ src/data/index.ts
@@ -4,6 +4,7 @@ export {
 	type RecordIdValue,
 	escape_ident,
 } from "./types/recordid.ts";
+
 export {
 	Range,
 	type Bound,
@@ -11,11 +12,7 @@ export {
 	BoundExcluded,
 	RecordIdRange,
 } from "./types/range.ts";
-export { Future } from "./types/future.ts";
-export { Uuid } from "./types/uuid.ts";
-export { Duration } from "./types/duration.ts";
-export { Decimal } from "./types/decimal.ts";
-export { Table } from "./types/table.ts";
+
 export {
 	Geometry,
 	GeometryCollection,
@@ -26,4 +23,11 @@ export {
 	GeometryPoint,
 	GeometryPolygon,
 } from "./types/geometry.ts";
+
+export { Future } from "./types/future.ts";
+export { Uuid } from "./types/uuid.ts";
+export { Duration } from "./types/duration.ts";
+export { Decimal } from "./types/decimal.ts";
+export { Table } from "./types/table.ts";
+
 export { encodeCbor, decodeCbor } from "./cbor.ts";
