+++ src/data/types/range.ts
@@ -1,7 +1,3 @@
-import { Tagged } from "../../cbor";
-import { SurrealDbError } from "../../errors";
-import { toSurrealqlString } from "../../util/to-surrealql-string";
-import { TAG_BOUND_EXCLUDED, TAG_BOUND_INCLUDED, TAG_RANGE } from "../cbor";
 import {
 	type RecordIdValue,
 	escape_id_part,
@@ -9,11 +5,16 @@ import {
 	isValidIdPart,
 } from "./recordid";
 
+import { Tagged } from "../../cbor";
+import { SurrealDbError } from "../../errors";
+import { toSurrealqlString } from "../../util/toSurrealqlString";
+import { TAG_BOUND_EXCLUDED, TAG_BOUND_INCLUDED } from "../cbor";
+
 export class Range<Beg, End> {
 	constructor(
 		readonly beg: Bound<Beg>,
 		readonly end: Bound<End>,
-	) {}
+	) { }
 
 	toJSON(): string {
 		return this.toString();
@@ -28,11 +29,11 @@ export class Range<Beg, End> {
 
 export type Bound<T> = BoundIncluded<T> | BoundExcluded<T> | undefined;
 export class BoundIncluded<T> {
-	constructor(readonly value: T) {}
+	constructor(readonly value: T) { }
 }
 
 export class BoundExcluded<T> {
-	constructor(readonly value: T) {}
+	constructor(readonly value: T) { }
 }
 
 export class RecordIdRange<Tb extends string = string> {
