+++ src/index.ts
@@ -1,6 +1,7 @@
 export { Emitter, type Listener, type UnknownEvents } from "./util/emitter.ts";
-export { surql, surrealql } from "./util/tagged-template.ts";
-export { PreparedQuery } from "./util/PreparedQuery.ts";
+export { surql, surrealql } from "./util/taggedTemplate.ts";
+export { PreparedQuery } from "./util/preparedQuery.ts";
+
 export * as cbor from "./cbor";
 export * from "./cbor/gap";
 export * from "./cbor/error";
@@ -10,8 +11,9 @@ export * from "./types.ts";
 export * from "./util/jsonify.ts";
 export * from "./util/versionCheck.ts";
 export * from "./util/getIncrementalID.ts";
-export * from "./util/string-prefixes.ts";
-export * from "./util/to-surrealql-string.ts";
+export * from "./util/stringPrefixes.ts";
+export * from "./util/toSurrealqlString.ts";
+
 export {
 	ConnectionStatus,
 	AbstractEngine,
@@ -19,4 +21,5 @@ export {
 	type Engines,
 	type EngineEvents,
 } from "./engines/abstract.ts";
+
 export { Surreal, Surreal as default } from "./surreal.ts";
