+++ src/surreal.ts
@@ -6,6 +6,7 @@ import {
 	decodeCbor,
 	encodeCbor,
 } from "./data";
+
 import {
 	type AbstractEngine,
 	ConnectionStatus,
@@ -13,10 +14,6 @@ import {
 	type EngineEvents,
 	type Engines,
 } from "./engines/abstract.ts";
-import { PreparedQuery } from "./util/PreparedQuery.ts";
-import { Emitter } from "./util/emitter.ts";
-import { processAuthVars } from "./util/processAuthVars.ts";
-import { versionCheck } from "./util/versionCheck.ts";
 
 import {
 	type AccessRecordAuth,
@@ -33,21 +30,24 @@ import {
 	convertAuth,
 } from "./types.ts";
 
-import { type Fill, partiallyEncodeObject } from "./cbor";
-import { replacer } from "./data/cbor.ts";
-import type { RecordIdRange } from "./data/types/range.ts";
-import { HttpEngine } from "./engines/http.ts";
-import { WebsocketEngine } from "./engines/ws.ts";
 import {
 	EngineDisconnected,
 	NoActiveSocket,
-	NoDatabaseSpecified,
-	NoNamespaceSpecified,
 	NoTokenReturned,
 	ResponseError,
 	SurrealDbError,
 	UnsupportedEngine,
 } from "./errors.ts";
+
+import { PreparedQuery } from "./util/preparedQuery.ts";
+import { Emitter } from "./util/emitter.ts";
+import { processAuthVars } from "./util/processAuthVars.ts";
+import { versionCheck } from "./util/versionCheck.ts";
+import { type Fill, partiallyEncodeObject } from "./cbor";
+import { replacer } from "./data/cbor.ts";
+import type { RecordIdRange } from "./data/types/range.ts";
+import { HttpEngine } from "./engines/http.ts";
+import { WebsocketEngine } from "./engines/ws.ts";
 
 type R = Prettify<Record<string, unknown>>;
 type RecordId<Tb extends string = string> = _RecordId<Tb> | StringRecordId;
@@ -425,12 +425,12 @@ export class Surreal {
 		const params =
 			q instanceof PreparedQuery
 				? [
-						q.query,
-						partiallyEncodeObject(q.bindings, {
-							fills: b as Fill[],
-							replacer: replacer.encode,
-						}),
-					]
+					q.query,
+					partiallyEncodeObject(q.bindings, {
+						fills: b as Fill[],
+						replacer: replacer.encode,
+					}),
+				]
 				: [q, b];
 
 		await this.ready;
