+++ src/util/getIncrementalID.ts
@@ -1,4 +1,5 @@
 let id = 0;
+
 export function getIncrementalID(): string {
 	id = (id + 1) % Number.MAX_SAFE_INTEGER;
 	return id.toString();
